refactor: NewCard 관련 코드 패키지 정리
