+++ app/src/main/java/nextstep/payments/ui/components/PaymentCard.kt
@@ -1,4 +1,4 @@
-package nextstep.payments
+package nextstep.payments.ui.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
