+++ app/src/main/java/nextstep/payments/ui/newcard/NewCardTopBar.kt
@@ -1,4 +1,4 @@
-package nextstep.payments
+package nextstep.payments.ui.newcard
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
